Patient Intake Form
Ratinder and Jonathan worked on code for the patient form class. This form asks for their contact information and the date they want to schedule an appointment. We also put in a calendar that lets patients pick a date for their appointment. And if a date is already booked, the calendar won't let anyone else choose that day for an appointment. After the patient fills out the form, when they press the save button, a message will be displayed depending on whether they've completed all the required fields or not.